#721 refactor: name changes to reflect shifted roles, added deprecated methods/class/interface for backward compatibility.
