+++ core/http/client/src/main/java/org/eclipse/rdf4j/http/client/SessionManagerDependent.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2017 Eclipse RDF4J contributors.
+ * Copyright (c) 2015 Eclipse RDF4J contributors, Aduna, and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Distribution License v1.0
  * which accompanies this distribution, and is available at
@@ -8,11 +8,13 @@
 package org.eclipse.rdf4j.http.client;
 
 /**
- * @author Jeen Broekstra
- * @deprecated use {@link SessionManagerDependent} instead.
+ * Common interface for objects, such as Repository and RepositoryConnection, that are dependent on
+ * {@link HttpClientSessionManager}.
+ * 
+ * @author James Leigh
  */
-@Deprecated
-public interface SesameClientDependent {
+@SuppressWarnings("deprecation")
+public interface SessionManagerDependent extends SesameClientDependent {
 
 	/**
 	 * {@link HttpClientSessionManager} that has been assigned or has been used by this object. The life cycle
@@ -24,14 +26,6 @@ public interface SesameClientDependent {
 	HttpClientSessionManager getHttpClientSessionManager();
 
 	/**
-	 * @deprecated use {@link #getHttpClientSessionManager()} instead.
-	 */
-	@Deprecated
-	default HttpClientSessionManager getSesameClient() {
-		return getHttpClientSessionManager();
-	}
-
-	/**
 	 * Assign an {@link HttpClientSessionManager} that this object should use. The life cycle of the given
 	 * {@link HttpClientSessionManager} is independent of this object. Closing or shutting down this object
 	 * does not have any impact on the given client. Callers must ensure that the given client is properly
@@ -41,11 +35,4 @@ public interface SesameClientDependent {
 	 */
 	void setHttpClientSessionManager(HttpClientSessionManager client);
 
-	/**
-	 * @deprecated use {@link #setHttpClientSessionManager()} instead.
-	 */
-	@Deprecated
-	default void setSesameClient(SesameClient client) {
-		setHttpClientSessionManager((HttpClientSessionManager)client);
-	}
 }
