+++ core/repository/sparql/src/main/java/org/eclipse/rdf4j/repository/sparql/query/SPARQLTupleQuery.java
@@ -9,7 +9,7 @@ package org.eclipse.rdf4j.repository.sparql.query;
 
 import java.io.IOException;
 
-import org.eclipse.rdf4j.http.client.SparqlSession;
+import org.eclipse.rdf4j.http.client.SPARQLProtocolSession;
 import org.eclipse.rdf4j.http.client.query.AbstractHTTPQuery;
 import org.eclipse.rdf4j.query.MalformedQueryException;
 import org.eclipse.rdf4j.query.QueryEvaluationException;
@@ -30,7 +30,7 @@ public class SPARQLTupleQuery extends AbstractHTTPQuery implements TupleQuery {
 	// TODO there was some magic going on in SparqlOperation to get baseURI
 	// directly replaced within the query using BASE
 
-	public SPARQLTupleQuery(SparqlSession httpClient, String baseUri, String queryString) {
+	public SPARQLTupleQuery(SPARQLProtocolSession httpClient, String baseUri, String queryString) {
 		super(httpClient, QueryLanguage.SPARQL, queryString, baseUri);
 	}
 
@@ -38,7 +38,7 @@ public class SPARQLTupleQuery extends AbstractHTTPQuery implements TupleQuery {
 		throws QueryEvaluationException
 	{
 
-		SparqlSession client = getHttpClient();
+		SPARQLProtocolSession client = getHttpClient();
 		try {
 			return client.sendTupleQuery(QueryLanguage.SPARQL, getQueryString(), baseURI, dataset,
 					getIncludeInferred(), getMaxExecutionTime(), getBindingsArray());
@@ -58,7 +58,7 @@ public class SPARQLTupleQuery extends AbstractHTTPQuery implements TupleQuery {
 		throws QueryEvaluationException, TupleQueryResultHandlerException
 	{
 
-		SparqlSession client = getHttpClient();
+		SPARQLProtocolSession client = getHttpClient();
 		try {
 			client.sendTupleQuery(QueryLanguage.SPARQL, getQueryString(), baseURI, dataset,
 					getIncludeInferred(), getMaxExecutionTime(), handler, getBindingsArray());
